Fix stack overflow with tail-end optimization
The `split()` function is designed to use recursion, but in certain
cases -- such as if the only differing data is near the end of a large
amount of data while using a debug build -- it can cause a stack
overflow, particularly with the second recursive split() call. The
reason this seems to happen in Rust and not the C version is because
Rust debug builds appear to use a significantly larger amount of stack
space per function call. As it is, this function is unusable when used
with very large files while building debug.

This change adds a tail-end optimization and significantly reduces
recursion. It's probably also more optimal.